+++ crates/pool/src/lib.rs
@@ -18,6 +18,7 @@ pub mod root;
 pub mod tx;
 pub mod validator;
 pub mod world_chain_tx;
+pub mod world_chain_validator;
 
 /// Type alias for World Chain transaction pool
 pub type WorldChainTransactionPool<
